Whitespace removal and removal of unused variable
Whitespace is problematic because looking at differences in lines, gets
really confusing if it is only a whitespace change.  Better to not have
them there in the first place.

Also, it indicates that you are careful, or care about, how your code
looks, and what it contains, even if you can't usually see it.

It also helps to reduce file sizes, network traffic, etc.